Yet more compiler warning fixes.
git-svn-id: http://svn.mythtv.org/svn/trunk@17379 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythfreemheg/ParseText.cpp
@@ -58,7 +58,7 @@ MHParseText::~MHParseText()
 // Get the next character.
 void MHParseText::GetNextChar()
 {
-    if (m_p >= m_data.size())
+    if ((int)m_p >= m_data.size())
         m_ch = EOF;
     else m_ch = m_data[m_p++];
 }
